% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app1DIA_shiny.R
\name{runApp1DIA}
\alias{runApp1DIA}
\title{Runs the Shiny DIA App}
\usage{
runApp1DIA(options = list())
}
\arguments{
\item{options}{A list of options passed to \code{shiny::shinyApp()}.}
}
\description{
Runs the Shiny DIA App
}
