% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app1MaxQuant.R
\name{runApp1MaxQuant}
\alias{runApp1MaxQuant}
\title{Runs the Shiny MaxQuant App}
\usage{
runApp1MaxQuant(options = list())
}
\arguments{
\item{options}{A list of options passed to \code{shiny::shinyApp()}.}
}
\description{
Runs the Shiny MaxQuant App
}
