% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app1Spectronaut.R
\name{runApp1Spectronaut}
\alias{runApp1Spectronaut}
\title{Runs the Shiny Spectronaut App}
\usage{
runApp1Spectronaut(options = list())
}
\arguments{
\item{options}{A list of options passed to \code{shiny::shinyApp()}.}
}
\description{
Runs the Shiny Spectronaut App
}
