% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app1TMT_shiny.R
\name{runApp1TMT}
\alias{runApp1TMT}
\title{Runs the Shiny TMT App}
\usage{
runApp1TMT(options = list())
}
\arguments{
\item{options}{A list of options passed to \code{shiny::shinyApp()}.}
}
\description{
Runs the Shiny TMT App
}
