% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app2_shiny.R
\name{runApp2}
\alias{runApp2}
\title{Runs the Shiny App2 Statistical Analysis}
\usage{
runApp2(options = list())
}
\arguments{
\item{options}{A list of options passed to \code{shiny::shinyApp()}.}
}
\description{
Runs the Shiny App2 Statistical Analysis
}
