% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_workflows.R
\name{ScriptApp1DIA}
\alias{ScriptApp1DIA}
\title{Processes a ScaffoldDIA Samples Report and outputs Limma-ready tables}
\usage{
ScriptApp1DIA(
  project_name,
  protein_file,
  samples_file,
  P1exclude,
  P1groupexclude,
  GlobalGroupX,
  GlobalGroupY,
  required_observations,
  normalization_option,
  log2_check,
  remove_0_values,
  normalize_to_reference,
  type
)
}
\description{
Processes a ScaffoldDIA Samples Report and outputs Limma-ready tables
}
