% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_workflows.R
\name{ScriptApp1PhosphoTMT}
\alias{ScriptApp1PhosphoTMT}
\title{Processes a MaxQuant PhosphoTMT data set and outputs Limma-ready tables}
\usage{
ScriptApp1PhosphoTMT(
  project_name,
  protein_file,
  samples_file,
  phospho_file,
  P1exclude,
  P1groupexclude,
  GlobalGroupX,
  GlobalGroupY,
  required_observations,
  normalization_option,
  log2_check,
  remove_0_values,
  normalize_to_reference,
  normalize_to_protein,
  type
)
}
\description{
Processes a MaxQuant PhosphoTMT data set and outputs Limma-ready tables
}
